+++ ldes-fragmentisers/ldes-fragmentisers-common/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentation/FragmentDeleterScheduler.java
@@ -1,14 +1,11 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.compaction.application.services;
+package be.vlaanderen.informatievlaanderen.ldes.server.fragmentation;
 
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.events.fragmentation.FragmentDeletedEvent;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.model.TreeRelation;
 import be.vlaanderen.informatievlaanderen.ldes.server.fragmentation.entities.Fragment;
 import be.vlaanderen.informatievlaanderen.ldes.server.fragmentation.repository.FragmentRepository;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class FragmentDeleterScheduler {
@@ -27,22 +24,10 @@ public class FragmentDeleterScheduler {
 				.getDeletionCandidates()
 				.filter(Fragment::isReadyForDeletion)
 				.forEach(readyForDeletionFragment -> {
-					fragmentRepository.deleteFragment(readyForDeletionFragment);
-					removeRelationsPointingToDeletedFragment(readyForDeletionFragment);
+					fragmentRepository.deleteFragmentAndRemoveRelationsPointingToFragment(readyForDeletionFragment);
 					applicationEventPublisher
 							.publishEvent(new FragmentDeletedEvent(readyForDeletionFragment.getFragmentId()));
 				});
 	}
 
-	private void removeRelationsPointingToDeletedFragment(Fragment readyForDeletionFragment) {
-		List<Fragment> fragments = fragmentRepository
-				.retrieveFragmentsByOutgoingRelation(readyForDeletionFragment.getFragmentId());
-		fragments.forEach(fragment -> {
-			List<TreeRelation> relationsToRemove = fragment.getRelations().stream()
-					.filter(treeRelation -> treeRelation.treeNode().equals(readyForDeletionFragment.getFragmentId()))
-					.toList();
-			relationsToRemove.forEach(fragment::removeRelation);
-			fragmentRepository.saveFragment(fragment);
-		});
-	}
 }
