+++ ldes-fragmentisers/ldes-fragmentisers-substring/src/test/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentisers/substring/fragment/SubstringFragmentFinderTest.java
@@ -1,7 +1,7 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.fragmentisers.substring.fragment;
 
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.model.LdesFragmentIdentifier;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.model.FragmentPair;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.model.LdesFragmentIdentifier;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.model.ViewName;
 import be.vlaanderen.informatievlaanderen.ldes.server.fragmentation.entities.Fragment;
 import be.vlaanderen.informatievlaanderen.ldes.server.fragmentisers.substring.config.SubstringConfig;
@@ -53,10 +53,10 @@ class SubstringFragmentFinderTest {
 	void when_RootFragmentHasReachedItsLimit_FirstOpenFragmentIsReturned() {
 		Fragment rootFragment = new Fragment(new LdesFragmentIdentifier(
 				VIEW_NAME, List.of(new FragmentPair(SUBSTRING, "\"\""))), false, 1,
-				List.of(), deleteTime);
+				List.of(), null);
 		Fragment aFragment = new Fragment(new LdesFragmentIdentifier(
 				VIEW_NAME, List.of(new FragmentPair(SUBSTRING, "a"))), false, 1,
-				List.of(), deleteTime);
+				List.of(), null);
 		Fragment abFragment = PARENT_FRAGMENT.createChild(new FragmentPair(SUBSTRING, "ab"));
 		when(substringFragmentCreator.getOrCreateSubstringFragment(PARENT_FRAGMENT,
 				"a")).thenReturn(aFragment);
